enhance precompute price service

+++ app/product/serializers/price_configuration.py
@@ -17,7 +17,7 @@ class ProductPriceConfigurationSerializer(serializers.ModelSerializer):
             return super().create(*args, **kwargs)
         except Exception as e:
             raise serializers.ValidationError(e)
-        
+
     def update(self, *args, **kwargs):
         try:
             return super().update(*args, **kwargs)
@@ -34,7 +34,7 @@ class ProductPricePreviewResponseSerializer(serializers.Serializer):
     product_name = serializers.CharField()
     price_configuration_id = serializers.IntegerField()
     price_configuration_name = serializers.CharField()
-    base_price_vnd = serializers.DecimalField(max_digits=10, decimal_places=2)
-    price_vnd = serializers.DecimalField(max_digits=10, decimal_places=2)
-    base_price_usd = serializers.DecimalField(max_digits=10, decimal_places=2)
-    price_usd = serializers.DecimalField(max_digits=10, decimal_places=2)
+    base_price_vnd = serializers.FloatField()
+    price_vnd = serializers.FloatField()
+    base_price_usd = serializers.FloatField()
+    price_usd = serializers.FloatField()
